Decompose schemas/manifest.py into schemas/manifest/ package
The 1094-line schemas/manifest.py absorbed every new manifest section as
the wedge widened past Tool-Use Readiness. Split it into one module per
top-level shipgate.yaml section so each concept lives in a predictable,
isolated location — making future additions a one-file change instead of
a big-file edit.

Layout (dependency-ordered): _common + _artifacts foundations; project /
agent / environment metadata; tool_sources + per-framework configs
(openai_api, anthropic, google_adk, langchain, crewai, codex_plugin,
n8n, validation); policy / permissions / risk_overrides governance;
policy_packs + severity_overrides + checks composition; action_surface
governance lens; baseline / ci / output build-time knobs; root.py for
the AgentsShipgateManifest aggregate. The package __init__.py re-exports
every public symbol so `from agents_shipgate.schemas.manifest import X`
keeps working unchanged for every consumer.

Verified zero behavior change: scripts/generate_schemas.py --check
exits 0 (docs/manifest-v0.1.json byte-identical), all 2106 tests pass,
ruff clean. No other file in the repo touched. Largest file after the
split is action_surface.py at 210 LOC; no file over 220 LOC.

Mirrors the v0.21 E8 inputs/n8n.py → inputs/n8n/ decomposition pattern.
