+++ src/agents_shipgate/schemas/manifest/__init__.py
@@ -1,0 +1,169 @@
+"""Manifest schema package.
+
+The schema for ``shipgate.yaml`` is decomposed into one module per
+top-level manifest section. This ``__init__`` re-exports every public
+symbol so ``from agents_shipgate.schemas.manifest import X`` keeps
+working for every X that used to live in the legacy monolithic
+``schemas/manifest.py`` module.
+
+If you are adding a new manifest section:
+
+1. Create ``schemas/manifest/<section>.py`` with the Pydantic models.
+2. Add the field on ``AgentsShipgateManifest`` in ``root.py``.
+3. Re-export the new models below (under the alphabetised section block).
+4. Regenerate ``docs/manifest-v0.1.json`` via
+   ``python scripts/generate_schemas.py``.
+
+Schema layering (enforced by ``tests/test_schema_boundaries.py``):
+modules in this package may import from ``schemas.common`` and from
+each other, but must NOT import from ``agents_shipgate.core.*``.
+"""
+
+from agents_shipgate.schemas.manifest._artifacts import (
+    ArtifactPathConfig,
+    NamedArtifactPathConfig,
+)
+from agents_shipgate.schemas.manifest.action_surface import (
+    ActionApprovalConfig,
+    ActionDeclarationConfig,
+    ActionEffect,
+    ActionEvidenceConfig,
+    ActionPolicyConfig,
+    ActionPolicyMatchConfig,
+    ActionRiskTag,
+    ActionSafeguardsConfig,
+    ActionSurfaceConfig,
+)
+from agents_shipgate.schemas.manifest.agent import (
+    AgentConfig,
+    AgentSdkConfig,
+)
+from agents_shipgate.schemas.manifest.anthropic import AnthropicConfig
+from agents_shipgate.schemas.manifest.baseline import (
+    BaselineConfig,
+    BaselineIntegrityMode,
+)
+from agents_shipgate.schemas.manifest.checks import (
+    ChecksConfig,
+    SuppressionConfig,
+)
+from agents_shipgate.schemas.manifest.ci import CiConfig
+from agents_shipgate.schemas.manifest.codex_plugin import (
+    CodexPluginMcpInventoryConfig,
+    CodexPluginsConfig,
+)
+from agents_shipgate.schemas.manifest.crewai import CrewAiConfig
+from agents_shipgate.schemas.manifest.environment import EnvironmentConfig
+from agents_shipgate.schemas.manifest.google_adk import GoogleAdkConfig
+from agents_shipgate.schemas.manifest.langchain import LangChainConfig
+from agents_shipgate.schemas.manifest.n8n import N8nConfig
+from agents_shipgate.schemas.manifest.openai_api import OpenAIApiConfig
+from agents_shipgate.schemas.manifest.output import (
+    OutputConfig,
+    PacketOutputConfig,
+)
+from agents_shipgate.schemas.manifest.permissions import PermissionsConfig
+from agents_shipgate.schemas.manifest.policies import (
+    PoliciesConfig,
+    PolicyToolEntry,
+)
+from agents_shipgate.schemas.manifest.policy_packs import PolicyPackConfig
+from agents_shipgate.schemas.manifest.project import ProjectConfig
+from agents_shipgate.schemas.manifest.risk_overrides import (
+    RiskOverridesConfig,
+    ToolRiskOverride,
+)
+from agents_shipgate.schemas.manifest.root import AgentsShipgateManifest
+from agents_shipgate.schemas.manifest.severity_overrides import (
+    OverrideAcknowledgement,
+    SeverityOverrideEntry,
+)
+from agents_shipgate.schemas.manifest.tool_sources import (
+    BUILTIN_PER_SCAN_ONLY_TOOL_SOURCE_TYPES,
+    BUILTIN_TOOL_SOURCE_TYPES,
+    ToolSourceConfig,
+)
+from agents_shipgate.schemas.manifest.validation import (
+    ValidationConfig,
+    ValidationEvidenceConfig,
+    ValidationRequiredEvidenceConfig,
+)
+
+__all__ = [
+    # tool_sources constants
+    "BUILTIN_PER_SCAN_ONLY_TOOL_SOURCE_TYPES",
+    "BUILTIN_TOOL_SOURCE_TYPES",
+    # action_surface
+    "ActionApprovalConfig",
+    "ActionDeclarationConfig",
+    "ActionEffect",
+    "ActionEvidenceConfig",
+    "ActionPolicyConfig",
+    "ActionPolicyMatchConfig",
+    "ActionRiskTag",
+    "ActionSafeguardsConfig",
+    "ActionSurfaceConfig",
+    # agent
+    "AgentConfig",
+    "AgentSdkConfig",
+    # root
+    "AgentsShipgateManifest",
+    # anthropic
+    "AnthropicConfig",
+    # _artifacts
+    "ArtifactPathConfig",
+    # baseline
+    "BaselineConfig",
+    "BaselineIntegrityMode",
+    # checks
+    "ChecksConfig",
+    # ci
+    "CiConfig",
+    # codex_plugin
+    "CodexPluginMcpInventoryConfig",
+    "CodexPluginsConfig",
+    # crewai
+    "CrewAiConfig",
+    # environment
+    "EnvironmentConfig",
+    # google_adk
+    "GoogleAdkConfig",
+    # langchain
+    "LangChainConfig",
+    # n8n
+    "N8nConfig",
+    # _artifacts (named)
+    "NamedArtifactPathConfig",
+    # openai_api
+    "OpenAIApiConfig",
+    # output
+    "OutputConfig",
+    # severity_overrides
+    "OverrideAcknowledgement",
+    # output (packet)
+    "PacketOutputConfig",
+    # permissions
+    "PermissionsConfig",
+    # policies
+    "PoliciesConfig",
+    # policy_packs
+    "PolicyPackConfig",
+    # policies (entry)
+    "PolicyToolEntry",
+    # project
+    "ProjectConfig",
+    # risk_overrides
+    "RiskOverridesConfig",
+    # severity_overrides
+    "SeverityOverrideEntry",
+    # checks (suppression)
+    "SuppressionConfig",
+    # tool_sources
+    "ToolSourceConfig",
+    # risk_overrides
+    "ToolRiskOverride",
+    # validation
+    "ValidationConfig",
+    "ValidationEvidenceConfig",
+    "ValidationRequiredEvidenceConfig",
+]
